new refactory intermediate test
